feat: implement E2E testing framework with WebdriverIO and Appium
- Added configuration files for E2E testing in the `test/e2e` directory.
- Created `HomeOnboardingScreen` page object for managing onboarding interactions.
- Developed initial test for wallet app onboarding, verifying UI elements and welcome messages.
- Updated `package.json` and `package-lock.json` to include necessary dependencies for testing.
- Modified `.gitignore` to exclude test artifacts and configuration files.
- Removed legacy E2E testing setup from `e2e` directory.
