--- e2e/package.json
@@ -1,24 +1,0 @@
-{
-  "name": "wdk-starter-react-native-e2e",
-  "version": "0.0.1",
-  "description": "WDK Starter React Native E2E Tests",
-  "scripts": {
-    "test": "export ANDROID_HOME=$HOME/Library/Android/sdk && export PATH=$PATH:$ANDROID_HOME/platform-tools && wdio run wdio.conf.ts",
-    "build-apk": "export JAVA_HOME=$(/usr/libexec/java_home -v 21) && export PATH=$JAVA_HOME/bin:$PATH && cd .. && npx expo export --platform android --output-dir ./dist && npx expo prebuild --platform android --clean && echo \"sdk.dir=$HOME/Library/Android/sdk\" > android/local.properties && cd android && ./gradlew assembleRelease && cp app/build/outputs/apk/release/app-release.apk ../e2e/apps/app-release.apk",
-    "build-ios": "cd ../ios && xcodebuild -workspace wdkstarterreactnative.xcworkspace -scheme wdkstarterreactnative -configuration Debug -sdk iphonesimulator -derivedDataPath build -quiet && find build -name '*.app' -exec cp -r {} ../e2e/apps/wdkstarterreactnative.app \\;",
-    "appium": "appium",
-    "check-devices": "adb devices",
-    "check-ios-devices": "xcrun simctl list devices"
-  },
-  "devDependencies": {
-    "@types/mocha": "^10.0.10",
-    "@types/node": "^20.0.0",
-    "@wdio/appium-service": "^9.13.0",
-    "@wdio/cli": "^9.13.0",
-    "@wdio/globals": "^9.13.0",
-    "@wdio/local-runner": "^9.13.0",
-    "@wdio/mocha-framework": "^9.13.0",
-    "@wdio/spec-reporter": "^9.13.0",
-    "typescript": "^5.0.0"
-  }
-}
